src/main/java/ru/job4j/tracker/EditAction.java
package ru.job4j.tracker;

public class EditAction implements UserAction {
    private final Output out;

    public EditAction(Output out) {
        this.out = out;
    }

    @Override
    public String name() {
        return "Edit item";
    }

    @Override
    public boolean execute(Input input, Store store) {
        out.println("=== Edit item ===");
        int id = input.askInt("Enter id ");
        String name = input.askStr("Enter name: ");
        Item item = new Item(name);
        if (store.replace(id, item)) {
            out.println("Edit item is done.");
        } else {
            out.println("Item not found.");
        }
        return true;
    }

}
